refactor: Move CandidateType to nvi-commons
