+++ nvi-commons/src/main/java/no/sikt/nva/nvi/common/dto/CandidateType.java
@@ -1,4 +1,4 @@
-package no.sikt.nva.nvi.events.model;
+package no.sikt.nva.nvi.common.dto;
 
 import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
@@ -12,6 +12,5 @@ import java.net.URI;
 })
 @JsonTypeInfo(use = JsonTypeInfo.Id.NAME, property = "type")
 public sealed interface CandidateType permits NonNviCandidate, NviCandidate {
-
   URI publicationId();
 }
